+++ R/betaDelta-beta-delta.R
@@ -11,12 +11,9 @@
 #'   \item{lm_process}{Pre-processed object of class `lm`.}
 #'   \item{type}{Standard error type.}
 #'   \item{gamma}{Asymptotic covariance matrix of the sample covariance matrix.}
-#'   \item{est}{Vector of standardized slopes.}
-#'   \item{vcov}{Sampling covariance matrix of the standardized slopes.}
 #'   \item{acov}{Asymptotic covariance matrix of the standardized slopes.}
-#'   \item{n}{Sample size.}
-#'   \item{p}{Number of regressors.}
-#'   \item{df}{\eqn{n - p - 1} degrees of freedom.}
+#'   \item{vcov}{Sampling covariance matrix of the standardized slopes.}
+#'   \item{est}{Vector of standardized slopes.}
 #' }
 #' @param object Object of class `lm`.
 #' @param type Character string.
@@ -37,12 +34,19 @@
 #' @examples
 #' object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
 #' std <- BetaDelta(object)
-#' # Methods -------------------------------------------------------
+#' # Methods ----------------------------------------------------
 #' print(std)
 #' summary(std)
 #' coef(std)
 #' vcov(std)
 #' confint(std, level = 0.95)
+#' ## Differences of standardized regression coefficients -------
+#' out <- dif(std)
+#' print(out)
+#' summary(out)
+#' coef(out)
+#' vcov(out)
+#' confint(out, level = 0.95)
 #' @export
 #' @family Beta Delta Functions
 #' @keywords betaDelta
@@ -86,9 +90,9 @@ BetaDelta <- function(object,
     lm_process = lm_process,
     type = type,
     gamma = gamma,
-    est = lm_process$betastar,
+    acov = acov,
     vcov = vcov,
-    acov = acov
+    est = lm_process$betastar
   )
   class(out) <- c(
     "betadelta",