+++ R/betaDelta-diff-beta-ci-dot.R
@@ -25,8 +25,7 @@
       se = sqrt(diag(object$vcov)),
       theta = 0,
       alpha = alpha,
-      z = FALSE,
-      df = object$betadelta$lm_process$df
+      z = TRUE
     )
   )
 }